refactor: move common used attach/ptrace function
